refactor datatype package structure
