--- lib/src/datatype.dart
@@ -1,6 +1,0 @@
-library datatype;
-
-export 'package:javelin/src/datatype/either.dart';
-export 'package:javelin/src/datatype/id.dart';
-export 'package:javelin/src/datatype/option.dart';
-export 'package:javelin/src/datatype/tuple.dart';
