+++ lib/src/datatype/core/extension/either.dart
@@ -1,4 +1,4 @@
-import 'package:javelin/src/datatype.dart';
+import 'package:javelin/src/datatype/datatype_core.dart';
 import 'package:javelin/src/typeclass.dart';
 
 //final eitherTypeInstance = EitherType<Object>();
