+++ lib/src/datatype/core/extension/option.dart
@@ -1,5 +1,5 @@
+import 'package:javelin/src/datatype/datatype_core.dart';
 import 'package:javelin/src/typeclass.dart';
-import 'package:javelin/src/datatype/option.dart';
 
 final optionTypeInstance = OptionType._();
 
@@ -34,17 +34,6 @@ class OptionType
     Kind<ForOption, B> f(A a),
   ) =>
       fa.fix().fold(() => Option.none(), f);
-}
-
-class Unit {
-  const Unit();
-
-  @override
-  String toString() => 'Unit';
-  @override
-  bool operator ==(other) => other is Unit;
-  @override
-  int get hashCode => 0;
 }
 
 /*
