+++ lib/src/datatype/core/tuple.dart
@@ -1,3 +1,5 @@
+part of datatype_core;
+
 class Tuple2<A, B> {
   final A a;
   final B b;
