+++ lib/src/datatype/datatype_core.dart
@@ -1,0 +1,11 @@
+library datatype_core;
+
+import 'package:javelin/src/datatype/core/extension/either.dart';
+import 'package:javelin/src/datatype/core/extension/id.dart';
+import 'package:javelin/src/datatype/core/extension/option.dart';
+import 'package:javelin/src/typeclass.dart';
+
+part 'core/either.dart';
+part 'core/id.dart';
+part 'core/option.dart';
+part 'core/tuple.dart';
