+++ test/datatype/core/either_test.dart
@@ -1,17 +1,17 @@
-import 'package:javelin/src/datatype.dart';
+import 'package:javelin/src/datatype/datatype_core.dart';
 import 'package:test/test.dart';
 
 import '../../gen.dart';
 import '../../law.dart';
-import '../laws/applicative_error_laws.dart';
-import '../laws/applicative_laws.dart';
-import '../laws/bifunctor_laws.dart';
-import '../laws/eq_laws.dart';
-import '../laws/functor_laws.dart';
-import '../laws/invariant_laws.dart';
-import '../laws/monad_error_laws.dart';
-import '../laws/monad_laws.dart';
-import '../laws/show_laws.dart';
+import '../../typeclass/laws/applicative_error_laws.dart';
+import '../../typeclass/laws/applicative_laws.dart';
+import '../../typeclass/laws/bifunctor_laws.dart';
+import '../../typeclass/laws/eq_laws.dart';
+import '../../typeclass/laws/functor_laws.dart';
+import '../../typeclass/laws/invariant_laws.dart';
+import '../../typeclass/laws/monad_error_laws.dart';
+import '../../typeclass/laws/monad_laws.dart';
+import '../../typeclass/laws/show_laws.dart';
 
 Iterable<Law> eitherLaws() sync* {
   yield* ShowLaws.laws(
