+++ test/gen.dart
@@ -2,9 +2,8 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:math';
 
+import 'package:javelin/src/datatype/datatype_core.dart';
 import 'package:javelin/src/typeclass.dart';
-import 'package:javelin/src/datatype/either.dart';
-import 'package:javelin/src/typeclass/instance/option.dart';
 
 abstract class Gen<T> {
   T random();
