+++ test/typeclass/laws/applicative_error_laws.dart
@@ -1,4 +1,4 @@
-import 'package:javelin/src/datatype.dart';
+import 'package:javelin/src/datatype/datatype_core.dart';
 import 'package:javelin/src/typeclass.dart';
 import 'package:javelin/src/core.dart';
